requirements.txt
langgraph
langchain
langchain-anthropic
langchain-community
langchain-text-splitters
chromadb
python-dotenv
pypdf
python-docx
python-pptx
openpyxl
streamlit
msal
requests